requirements.txt
albumentations==1.4.13
black==24.8.0
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
pandas==2.2.2
pillow==10.4.0
pydantic==2.8.2
requests==2.32.3
rich==13.7.1
seaborn==0.13.2
segmentation-models-pytorch==0.3.3
tornado==6.4.1
tqdm==4.66.4
csv-logger==1.3.0
